Merged Edit Track Window into Main Screen
Merged the Edit Track window into the Main Screen. It will now display
the Gap- and Wishlist, as well as the votes. It's still untested and
won't update the Wishlist-Table, when the Wishlist was edited. Will
implement that as soon as I am able to test the current state of
Implementation.